// src/app/api-client/model/userFilterModel.ts
/**
 * WebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { UserFilter } from './userFilter';

export interface UserFilterModel { 
    userFilter?: UserFilter;
    currentPage?: number;
    pageSize?: number;
}